+++ lib/index.js
@@ -6,7 +6,7 @@ var AsyncEventEmitter = require('async-eventemitter')
 // require the percomiled contracts
 const num01 = require('./precompiled/01-ecrecover.js')
 const num02 = require('./precompiled/02-sha256.js')
-const num03 = require('./precompiled/03-repemd160.js')
+const num03 = require('./precompiled/03-ripemd160.js')
 const num04 = require('./precompiled/04-identity.js')
 
 module.exports = VM
